# Optimize canonicalizeContentType