--- a/protocol_test.go
+++ b/protocol_test.go
@@ -29,7 +29,7 @@ func TestCanonicalizeContentType(t *testing.T) {
 	}{
 		{name: "charset param should be treated as lowercase", arg: "application/json; charset=UTF-8", want: "application/json; charset=utf-8"},
 		{name: "non charset param should not be changed", arg: "multipart/form-data; boundary=fooBar", want: "multipart/form-data; boundary=fooBar"},
-		{name: "no parameters should not have base normalized", arg: "APPLICATION/json;  ", want: "application/json"},
+		{name: "no parameters should be normalized", arg: "APPLICATION/json;  ", want: "application/json"},
 	}
 	for _, tt := range tests {
 		tt := tt