@mattrobenolt 2023-01-26
As a follow up, maybe a micro-optimization, but it might be worth having a switch statement to check for the known types to avoid even going through the `mime.ParseMediaType` path.

I only think this is a worthwhile optimization to make because for a connect client and a connect server, which in our case is all the time, we know the types it sent within a set of known types. So it seems pretty valid IMO to have a fast path for known client <> server hot path.