@mattrobenolt 2023-01-26
Gross, but kinda what I came up with:

```diff
diff --git a/protocol.go b/protocol.go
index 3fd4369..2ba638f 100644
--- a/protocol.go
+++ b/protocol.go
@@ -316,6 +316,10 @@ func flushResponseWriter(w http.ResponseWriter) {
 }
 
 func canonicalizeContentType(ct string) string {
+	switch ct {
+	case "application/connect+proto", "application/connect+json", "application/json", "application/proto", "application/grpc", "application/grpc-web", "application/grpc+json", "application/grpc+proto", "application/grpc-web+json", "application/json; charset=utf-8", "application/json; charset=UTF-8":
+		return ct
+	}
 	base, params, err := mime.ParseMediaType(ct)
 	if err != nil {
 		return ct
```

I'm not sure if there's a simpler way to build this list, but this makes this function effectively free when paired with a connect-go client.

In go1.19+, I believe this switch gets optimized to a jump table as well.