@akshayjshah 2023-01-27
@mattrobenolt Spiritually, I'm on board with being more even aggressive here - as long as we're in here optimizing, let's create a fast path that avoids all use of `mime`. The hard-coded switch you're suggesting isn't _totally_ out of the question, but I'd like to avoid baking so much knowledge of the protocol(s) and encodings into this function. I think we can do nearly as well if we leverage a little knowledge of the mime-type grammar.

Let me push a commit here and see what you think.