@mattrobenolt 2023-01-27
I like this. Pretty solid middleground.