@akshayjshah 2023-01-27
@mattrobenolt LMK how this looks to you - with the commits I've pushed, the fast path is down to zero allocs and ~10ns/op. I think this is an okay middle ground - it gets a little into the weeds of what a valid content-type is, but it doesn't directly couple the `canonicalizeContentType` function to gRPC or Connect protocol details.