@mattrobenolt 2023-01-27
@akshayjshah I updated the benchmark in the description with the new hot path. Not quite down to 2ns, but I'll take 12ns. :)